feat: 接口校验完成

+++ packages/service/src/app/core/entity/base.entity.ts
@@ -10,7 +10,7 @@ import {
 export abstract class BaseEntity {
   @Column()
   @PrimaryGeneratedColumn({
-    type: 'bigint',
+    type: 'int',
     unsigned: true,
   })
   id: number;
@@ -23,24 +23,4 @@ export abstract class BaseEntity {
 
   @Column({ type: 'tinyint', nullable: true, default: 1 })
   status: number;
-
-  //   @CreateUserColumn()
-  //   @Column({ type: 'bigint', unsigned: true, transformer: idTransformer })
-  //   createdById?: number;
-
-  //   @UpdateUserColumn()
-  //   @Column({
-  //     type: 'bigint',
-  //     unsigned: true,
-  //     nullable: true,
-  //     transformer: idTransformer,
-  //   })
-  //   updatedById?: number;
-
-  //   @SoftDeleteColumn(IsDeleteEnum.TRUE)
-  //   @Column({ type: 'tinyint', unsigned: true, default: IsDeleteEnum.FALSE })
-  //   isDelete?: IsDeleteEnum;
-
-  //   @Column({ type: 'tinyint', unsigned: true, default: IsDeleteEnum.FALSE })
-  //   isDelete?: IsDeleteEnum;
 }
